refactor(trpc): relocate HTTPTRPCContext + context factory out of init.ts
Moves HTTPTRPCContext interface to trpc/types.ts (alongside BaseTRPCContext
it extends) and createHTTPTRPCContext factory to trpc/lib/create-http-context.ts.

init.ts now contains only tRPC instance creation + procedure definitions,
re-exporting createHTTPTRPCContext for backward-compatible imports.

Issue #193.
